fix: seperate hooks from utils directory of supabase
